avoid the gap counter overflow

+++ cooke.hpp
@@ -40,7 +40,7 @@ class CookeCamera: public Camera {
    PixelValue* pBlackImage;
 
    long firstFrameCounter; //first first frame's counter value
-   long totalGap;
+   long long totalGap;
 
    long nAcquiredFrames;
    //lock used to coordinate accessing to nAcquiredFrames & pLiveImage
